requirements.txt
torch>=2.0.0
torchvision>=0.16.0
Pillow>=9.5.0
tqdm>=4.65.0
